S8/pyproject.toml
[project]
name = "s7"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.38.18",
    "dotenv>=0.9.9",
    "faiss-cpu>=1.10.0",
    "fastapi>=0.115.12",
    "google-genai>=1.9.0",
    "llama-index>=0.12.28",
    "llama-index-embeddings-google-genai>=0.1.0",
    "markitdown[all]>=0.1.1",
    "mcp[cli]>=1.6.0",
    "paddleocr>=2.10.0",
    "paddlepaddle>=3.0.0",
    "pillow>=11.1.0",
    "pymupdf4llm>=0.0.24",
    "pypdf2>=3.0.1",
    "python-multipart>=0.0.20",
    "pyyaml>=6.0.2",
    "rich>=14.0.0",
    "scipy>=1.15.2",
    "tqdm>=4.67.1",
    "trafilatura>=2.0.0",
]
